+++ app-ios/iosApp/iosApp/iOSApp.swift
@@ -2,9 +2,12 @@ import SwiftUI
 
 @main
 struct iOSApp: App {
+    @StateObject private var audioRouteMonitor = AudioRouteMonitor()
+
     var body: some Scene {
         WindowGroup {
             ContentView()
+                .environmentObject(audioRouteMonitor)
         }
     }
 }